requirements.txt
#OpenAI
openai>=1.3.0

# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

#Langchain
langchain>=0.0.350
langchain-openai>=0.0.1
langchain-community>=0.0.1
